+++ src/application/UserApplicationService.ts
@@ -1,5 +1,5 @@
 import { IUserRepository } from '../model/User/IUserRepository';
-import UserRepository, { UserDataFull } from '../repository/UserRepository';
+import UserRepository, { UserDataFull } from '../infrastructure/UserRepository';
 
 export class UserApplicationService {
   static readonly userRepository: IUserRepository = new UserRepository();
