+++ src/infrastructure/ITweetDataModel.ts
